+++ enums/createBrainsFields.ts
@@ -1,5 +1,6 @@
 export enum CreateBrainsFields {
 	NAME = "name",
-	CONTEXT = "context",
 	MODEL = "model",
+	REASONING = "reasoning",
+	SYSTEM = "system", // This is the context of your brain which lives in the system role of the conversation array of the brain.
 }