+++ islands/BrainContextFormInput.tsx
@@ -3,7 +3,7 @@ import { CreateBrainsFields } from "../enums/createBrainsFields.ts";
 export function BrainContextFormInput() {
 	return (
 		<>
-			<label for="context" class="mt-10 block">GIVE YOUR BRAIN A CONTEXT</label>
+			<label for={CreateBrainsFields.SYSTEM} class="mt-10 block">GIVE YOUR BRAIN A CONTEXT</label>
 			<textarea
 				onKeyDown={(e) => {
 					if (e.key === "Enter" && !e.shiftKey) {
@@ -12,8 +12,8 @@ export function BrainContextFormInput() {
 					}
 				}}
 				class="bg-white font-quantico text-black text-base w-104 pl-2 rounded-lg"
-				name={CreateBrainsFields.CONTEXT}
-				id={CreateBrainsFields.CONTEXT}
+				name={CreateBrainsFields.SYSTEM}
+				id={CreateBrainsFields.SYSTEM}
 				placeholder="Your Brain Context Goes Here. Be Thorough."
 				required
 			/>