+++ routes/feedBrain.tsx
@@ -3,7 +3,7 @@ import { FileMap } from "../enums/fileMap.ts";
 import { SiteMap } from "../enums/siteMap.ts";
 import { UserBrains } from "../types/brain.ts";
 import { define } from "../utils.ts";
-import brainFoodFiles from "../utils/server/filesList.ts";
+import brainFoodList from "../utils/server/brainFoodList.ts";
 
 export default define.page(async function feedBrain(ctx) {
 	const userBrains: UserBrains = JSON.parse( await Deno.readTextFile(FileMap.BRAIN) );
@@ -14,7 +14,17 @@ export default define.page(async function feedBrain(ctx) {
 	}
 
 	const brain = userBrains[brainName];
-	const brainFood = await brainFoodFiles();
+	const brainFood = await brainFoodList();
+	console.log(brain);
 
-	return <p>{userBrains[brainName].model}</p>
+	return (
+		<div class="flex flex-col items-center justify-center h-dvh">
+			<div class="flex flex-col items-start">
+				{
+					brainFood.map((x,y) => 
+						<p key={y} class="text-amber-500 font-cherrybomb text-2xl">{x.charAt(0).toUpperCase() + x.slice(1)}</p>
+				)}
+			</div>
+		</div>
+	);
 });