+++ CHANGELOG.md
@@ -51,3 +51,5 @@
 - **v-0.3.0**
     - Refactored selectBrain.tsx to a component which renders links that use url searchParams instead of cookies.
     - Removed components no longer necessary.
+    - Added reasoning radio buttons to CreateBrainForm.tsx component
+    - Added BrainReasoningFormInput.tsx to components.